refactor: dto 패키지 정리
