+++ server/appteam/src/main/java/com/pknuErrand/appteam/dto/errand/ErrandPaginationRequestVo.java
@@ -1,4 +1,4 @@
-package com.pknuErrand.appteam.dto.errand.getDto;
+package com.pknuErrand.appteam.dto.errand;
 
 import com.pknuErrand.appteam.Enum.Status;
 import lombok.AllArgsConstructor;
@@ -6,11 +6,9 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class ErrandDistancePaginationRequestVo {
+public class ErrandPaginationRequestVo {
     private Long pk;
-    private double latitude;
-    private double longitude;
+    private String cursor;
     private int limit;
     private Status status;
 }
-
