+++ server/appteam/src/main/java/com/pknuErrand/appteam/service/errand/ErrandService.java
@@ -3,11 +3,11 @@ package com.pknuErrand.appteam.service.errand;
 import com.pknuErrand.appteam.domain.errand.Errand;
 import com.pknuErrand.appteam.domain.errand.ErrandBuilder;
 import com.pknuErrand.appteam.Enum.Status;
-import com.pknuErrand.appteam.dto.errand.getDto.ErrandDistancePaginationRequestVo;
-import com.pknuErrand.appteam.dto.errand.getDto.ErrandListResponseDto;
-import com.pknuErrand.appteam.dto.errand.getDto.ErrandDetailResponseDto;
-import com.pknuErrand.appteam.dto.errand.getDto.ErrandPaginationRequestVo;
-import com.pknuErrand.appteam.dto.errand.saveDto.ErrandSaveRequestDto;
+import com.pknuErrand.appteam.dto.errand.ErrandDistancePaginationRequestVo;
+import com.pknuErrand.appteam.dto.errand.ErrandListResponseDto;
+import com.pknuErrand.appteam.dto.errand.ErrandDetailResponseDto;
+import com.pknuErrand.appteam.dto.errand.ErrandPaginationRequestVo;
+import com.pknuErrand.appteam.dto.errand.ErrandSaveRequestDto;
 import com.pknuErrand.appteam.domain.member.Member;
 import com.pknuErrand.appteam.dto.member.MemberErrandDto;
 import com.pknuErrand.appteam.exception.CustomException;
